Add artboard component
Artboard is rendered as a simple table with rows and cells, which we have to key
with sibling-unique identifiers as React requires us to do (don't ask). We also
get to use rum/cursor-in to track changes in nested structures like maps of maps
or vectors of vectors (this case). Reacting to a cursor narrows the number of
components that need to re-render. Very nice.

Design again follows simple steps:
- structure your component,
- identify interesting state (potentially narrow it to just right with cursors),
- add reactions to state,
- mount.