Add keys component
Shows off how to set React keys on components. React requires that a key is set
on sub-components that are part of an array or generator-produced. Other than
that it gives components stable identity I have no idea why this is important
and whether it should or even could be used in user code.