Add custom-props component
Shows that you could set custom properties on React's native component. Can't
imagine why I would ever want to do it this way, though. Perhaps JS interop
might sometime warrant this?

Interesting bits are all JS interop: (this-as ...) and (aget ...). One should
remember that once compiled we're in Javascript world, so "this" and friends are
all there to mess with.